requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-arango>=7.5.0
minio>=7.1.0
python-multipart>=0.0.6
httpx>=0.24.0
aiokafka>=0.12.0  # Para cuando implementes el consumidor aiokafka 0.12.0
minio>=7.2.20
httpx>=0.28.1